refactor(pages): refactoring profile page
Setting the right folder and componentizing elements
